requirements-tests.txt
microsoft-bonsai-api==0.1.3
python-dotenv==0.13.0
numpy==1.19.5
pandas==1.2.4
matplotlib==3.4.1
scikit-learn==0.23.2
xgboost==1.3.1
lightgbm==3.1.0
ray==1.0.0
tune-sklearn==0.2.1
pytorch==1.8.0
# pytorch-lightning==1.2.6
skorch==0.9.0
hyperopt==0.2.5
optuna==2.6.0
scikit-optimize==0.8.1
mlflow==1.15.0
rich==9.13.0
hydra-core==1.0.5
natsort==7.1.0